+++ pydantic/json.py
@@ -4,9 +4,7 @@ from enum import Enum
 from types import GeneratorType
 from uuid import UUID
 
-from .main import BaseModel
-
-__all__ = ['pydantic_encoder']
+__all__ = 'pydantic_encoder', 'custom_pydantic_encoder', 'timedelta_isoformat'
 
 
 def isoformat(o):
@@ -18,6 +16,7 @@ ENCODERS_BY_TYPE = {
     datetime.datetime: isoformat,
     datetime.date: isoformat,
     datetime.time: isoformat,
+    datetime.timedelta: lambda td: td.total_seconds(),
     set: list,
     frozenset: list,
     GeneratorType: list,
@@ -27,6 +26,7 @@ ENCODERS_BY_TYPE = {
 
 
 def pydantic_encoder(obj):
+    from .main import BaseModel
     if isinstance(obj, BaseModel):
         return obj.dict()
     elif isinstance(obj, Enum):
@@ -38,3 +38,20 @@ def pydantic_encoder(obj):
         raise TypeError(f"Object of type '{obj.__class__.__name__}' is not JSON serializable")
     else:
         return encoder(obj)
+
+
+def custom_pydantic_encoder(type_encoders, obj):
+    encoder = type_encoders.get(type(obj))
+    if encoder:
+        return encoder(obj)
+    else:
+        return pydantic_encoder(obj)
+
+
+def timedelta_isoformat(td: datetime.timedelta) -> str:
+    """
+    ISO 8601 encoding for timedeltas.
+    """
+    minutes, seconds = divmod(td.seconds, 60)
+    hours, minutes = divmod(minutes, 60)
+    return f'P{td.days}DT{hours:d}H{minutes:d}M{seconds:d}.{td.microseconds:06d}S'